- split 5but from 5bitlh

+++ strstrProfiling/strstrAVX2_V2.c
@@ -177,5 +177,5 @@ _noinline_ void Run_strstr_AVX2_V2_test()
 	}
 	double runtimeSec = EndTimer();
 	printf(" ... Done\n");
-	printf("Searches made %zd. Found the string %d times. Seconds : %f\n\n", searchesMade, (int)foundCount, (float)runtimeSec);
+	printf("Searches made %zu. Found the string %zu times. Seconds : %f\n\n", searchesMade, foundCount, (float)runtimeSec);
 }
